Do a more TransferPlugin-y check in AggregatorCore on the senderSlot returned from recvFragment. Allow users to enable a subset of defined sources/destinations using a FHiCL array.

+++ artdaq/TransferPlugins/TransferInterface.cc
@@ -1,6 +1,6 @@
 #include "artdaq/TransferPlugins/TransferInterface.hh"
 
-int artdaq::TransferInterface::my_rank = 0;
+size_t artdaq::TransferInterface::my_rank = 0;
 
 artdaq::TransferInterface::TransferInterface(const fhicl::ParameterSet& ps, Role role)
   : role_(role)
